+++ docs/CHANGELOG.md
@@ -4,7 +4,8 @@
 
 - buttons in the doc repo are a bit bigger for better hit targets
 
-- multi-column selection by checkbox now to make it a bit more obvious for users.
+- multi-column selection by checkbox now to make it a bit more obvious for 
+  users to select multiple documents
 
 - Clicking flagged or archived buttons no longer select the row which was 
   confusing.